Optimization of pit box signaling and towing
